grid07-cognitive-routing-rag-fullstack/backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
langchain>=0.2.0,<0.3.0
langgraph>=0.1.0,<0.2.0
langchain-openai>=0.1.0,<0.2.0
langchain-community>=0.2.0,<0.3.0
faiss-cpu>=1.8.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.7.0,<3.0.0
numpy>=1.26.0,<2.0.0
